vmitrait: move IsVFIOVMI and its dependencies from pkg/util
Move IsVFIOVMI and the predicates it depends on to pkg/vmitrait so all
VMI trait checks live in one place [1].
Update all call sites and remove the tests that exercised the helpers
as same-package unexported functions.

[1] https://github.com/kubevirt/kubevirt/issues/17156
