+++ pkg/util/util.go
@@ -41,14 +41,6 @@ const (
 	ENV_VAR_VIRT_LAUNCHER_LOG_VERBOSITY = "VIRT_LAUNCHER_LOG_VERBOSITY"
 )
 
-// Check if a VMI spec requests GPU
-func IsGPUVMI(vmi *v1.VirtualMachineInstance) bool {
-	if vmi.Spec.Domain.Devices.GPUs != nil && len(vmi.Spec.Domain.Devices.GPUs) != 0 {
-		return true
-	}
-	return false
-}
-
 // Check if a VMI spec requests VirtIO-FS
 func IsVMIVirtiofsEnabled(vmi *v1.VirtualMachineInstance) bool {
 	if vmi.Spec.Domain.Devices.Filesystems != nil {
@@ -59,19 +51,6 @@ func IsVMIVirtiofsEnabled(vmi *v1.VirtualMachineInstance) bool {
 		}
 	}
 	return false
-}
-
-// Check if a VMI spec requests a HostDevice
-func IsHostDevVMI(vmi *v1.VirtualMachineInstance) bool {
-	if vmi.Spec.Domain.Devices.HostDevices != nil && len(vmi.Spec.Domain.Devices.HostDevices) != 0 {
-		return true
-	}
-	return false
-}
-
-// Check if a VMI spec requests a VFIO device
-func IsVFIOVMI(vmi *v1.VirtualMachineInstance) bool {
-	return CountVFIODevices(vmi) > 0
 }
 
 func CountVFIODevices(vmi *v1.VirtualMachineInstance) int {
