+++ pkg/util/util_test.go
@@ -30,78 +30,6 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 )
 
-var _ = Describe("GPU VMI Predicates", func() {
-	It("should detect VMI with GPU device plugins", func() {
-		vmi := &v1.VirtualMachineInstance{
-			Spec: v1.VirtualMachineInstanceSpec{
-				Domain: v1.DomainSpec{
-					Devices: v1.Devices{
-						GPUs: []v1.GPU{
-							{
-								Name:       "gpu1",
-								DeviceName: "nvidia.com/gpu",
-							},
-						},
-					},
-				},
-			},
-		}
-		Expect(IsGPUVMI(vmi)).To(BeTrue())
-	})
-
-	It("should not detect GPU VMI when no GPUs are specified", func() {
-		vmi := &v1.VirtualMachineInstance{
-			Spec: v1.VirtualMachineInstanceSpec{
-				Domain: v1.DomainSpec{
-					Devices: v1.Devices{},
-				},
-			},
-		}
-		Expect(IsGPUVMI(vmi)).To(BeFalse())
-	})
-})
-
-var _ = Describe("Host Device VMI Predicates", func() {
-	It("should detect VMI with host device plugins", func() {
-		vmi := &v1.VirtualMachineInstance{
-			Spec: v1.VirtualMachineInstanceSpec{
-				Domain: v1.DomainSpec{
-					Devices: v1.Devices{
-						HostDevices: []v1.HostDevice{
-							{
-								Name:       "hostdev1",
-								DeviceName: "vendor.com/device",
-							},
-						},
-					},
-				},
-			},
-		}
-		Expect(IsHostDevVMI(vmi)).To(BeTrue())
-	})
-
-	It("should detect VMI with host device DRA", func() {
-		vmi := &v1.VirtualMachineInstance{
-			Spec: v1.VirtualMachineInstanceSpec{
-				Domain: v1.DomainSpec{
-					Devices: v1.Devices{
-						HostDevices: []v1.HostDevice{
-							{
-								Name: "hostdev1",
-								ClaimRequest: &v1.ClaimRequest{
-									ClaimName:   "hostdev-claim",
-									RequestName: "hostdev-request",
-								},
-							},
-						},
-					},
-				},
-			},
-		}
-		Expect(IsHostDevVMI(vmi)).To(BeTrue())
-	})
-})
-
 var _ = Describe("CountVFIODevices", func() {
 	DescribeTable("should count VFIO devices correctly",
 		func(devices v1.Devices, expected int) {
